backend/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
decorator==5.1.1
distro==1.9.0
Django==5.0.6
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kombu==5.3.7
numpy==1.26.4
openai==1.33.0
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
py==1.11.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-redis-lock==4.0.0
pytz==2024.1
redis==5.0.5
retry==0.9.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
python-pptx==0.6.23
